refactor: convert server.py to server package structure
• Move server.py → server/server.py with Server class and core functionality
• Move auth.py → server/auth.py to consolidate authentication within server package
• Create server/__init__.py exposing only run() function for public API
• Update runner.py imports to use new Server location
• Fix all test imports and patch decorators for new module structure
• Maintain backward compatibility through package-level exports
