+++ src/itential_mcp/server/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2025 Itential, Inc
+# GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
+
+from .server import run
+
+__all__ = ("run",)
